+++ src/WebhookManager.php
@@ -196,11 +196,12 @@ class WebhookManager
             'name' => $name,
             'class' => $webhook->class,
             'event' => $webhook->event,
-            'type' => $webhook->type,
+            'method' => $webhook->method,
             'url' => $webhook->url,
             'userAttributes' => $webhook->userAttributes,
             'senderAttributes' => $webhook->senderAttributes,
             'eventAttributes' => $webhook->eventAttributes,
+            'payloadTemplate' => $webhook->payloadTemplate
         ];
 
         if ($webhook->id) {
@@ -235,7 +236,7 @@ class WebhookManager
     private function _createWebhookQuery(): Query
     {
         return (new Query())
-            ->select(['id', 'groupId', 'enabled', 'name', 'class', 'event', 'type', 'url', 'userAttributes', 'senderAttributes', 'eventAttributes'])
+            ->select(['id', 'groupId', 'enabled', 'name', 'class', 'event', 'method', 'url', 'userAttributes', 'senderAttributes', 'eventAttributes', 'payloadTemplate'])
             ->from(['{{%webhooks}}']);
     }